Make functions as external in Gravity.sol and use calldata where possible
This makes all functions not called within the contract external.
External functions can use calldata directly which might be a gas
improvement, but in this case that's only realized for deployERC20 which
has enough free space on the stack to use calldata. Our other major
functions are too tight on stack space to use anything other than
memory.