WIP: use Gravity.sol ABI directly
Starting the process of directly using the ABI by removing the custom
parsing for ValsetUpdatedEvent and using functionality from ethers
abigen instead. Debating whether or not to leave it as a macro (which
makes it kind of difficult to introspect) or to check in a version of
the generated code.

Also moved the downcasting functions to the generic gravity_utils
package so other code can easily pull it in.