Complete conversion of ethereum_events.rs
All event types are now using abigen versions. Additionally, it appears
that the conversion of LogicCallExecutedEvent from a log was just marked
as "unimplemented" but now has been added as well. Also rewrote the
event nonce filtering using iterators and adaptors. There's probably a
way of further using traits to reduce code duplication here but it's not
a huge issue that it's not super DRY right now.