WIP: further fixes
Annoyingly, the Signer trait requires that the sign operation be async,
but using a LocalWallet as we are means that it can't throw any errors,
so returning a Result is kinda useless. However, because we must use
await, the message builder functions must also be async, so we have to
call async functions even though they don't return a Result.