Conversion of variable length u8s to fixed 32
There are several cases where we may want to convert a Vec<u8> or &[u8]
to a [u8; 32], so add some generic functions for that and remove the
specific case for invalidation ids. Additionally, during testing, it
appears that the recovery of the correct address during valset updates
was not working correctly. The Signature object of ethers can accept
different types of inputs, and if it's a vec of u8s rather than a
[u8; 32] it will hash the message again. We must convert out signature
explicitly into a [u8; 32] as it is already a hashed value in order to
get the correct address back from recover.