ethers-rs -> master, update abigen, bump deps
A required bugfix for contracts to function correctly is on ethers-rs
master. Here we depend on master and lock to that commit. Also have
re-run the ABI build with the new version of ethers, as well as fixed up
some use statements and type errors introduced in the newer version.
Additionally, we had to bump versions on actix and futures in order to
maintain compatibility with the latest version of ethers. We're
explicitly using std::result::Result because the new ethers prelude
includes a published type that is also called Result.