Fix U256 parse attempts from Strings
U256 appears to expect hex input by default when parsing a string. In
the few cases where we're taking in a string to convert to U256, use
from_dec_str explicitly. The particular usage in eth.rs probably didn't
work originally as described in the expect comment anyway, but it also
appears to be dead code, so just fixing it up to be more sane. If a
float was expected a bunch of different conversion would be necessary,
and it would also be losing information anyway.